requirements.txt
azure-functions
openai
azure-storage-blob 
azure-identity